# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restomate API",
        "description": "API de réservation de tables de restaurant. Gérez restaurants, tables et réservations avec authentification Sanctum.",
        "contact": {
            "name": "Alex AGO",
            "email": "alexyessougnonago@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://restomate-api.test/api/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/restaurants": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Liste tous les restaurants",
                "description": "Retourne la liste paginée des restaurants actifs",
                "operationId": "c2732fde08b6fb1b998768e0212d072a",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Cotonou"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Chez Mama"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste récupérée"
                    }
                }
            },
            "post": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Créer un restaurant (Admin)",
                "operationId": "047e8f41ad48124b86b8af32562c6353",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "city",
                                    "opening_time",
                                    "closing_time"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Chez Mama Africa"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Restaurant africain"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rue des Cocotiers"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Cotonou"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+22997000000"
                                    },
                                    "opening_time": {
                                        "type": "string",
                                        "example": "08:00"
                                    },
                                    "closing_time": {
                                        "type": "string",
                                        "example": "23:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Restaurant créé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/restaurants/{id}": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Détail d'un restaurant",
                "operationId": "fda112636f3f0744367f22ebf61831fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant récupéré"
                    },
                    "404": {
                        "description": "Introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Modifier un restaurant (Admin)",
                "operationId": "9d8efb7a86aeee93cfc1b54dd9f56763",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nouveau nom"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Porto-Novo"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restaurant mis à jour"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Archiver un restaurant (Admin)",
                "operationId": "992e6ded6ad1babc7467060e4f35bbdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant archivé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entre ton token Sanctum obtenu après login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}